+++ proxysiphon/__init__.py
@@ -1,3 +1,4 @@
 from proxysiphon.records import read_ncdc, read_lgm, read_petm
 from proxysiphon.agemodel import get_deltar_online, fit_agedepthmodel, date_proxy
-from . import qcreport
+from proxysiphon.lmr_hdf5 import nc2lmrh5, nc2lmrdf
+from proxysiphon import qcreport